all three channels run independently in python script. adc is not triggered by interruped any longer

+++ app/src/signal_out.c
@@ -112,8 +112,8 @@ void signal_timer1_ovf_isr(void)
         
     }
     OCR1A = sine_wave[sine_index[0]];
-    OCR1B = sine_wave[sine_index[0] - 42u ];
-    OCR2A = sine_wave[sine_index[0] - 84u];
+    OCR1B = sine_wave[(uint8_t)(sine_index[0] - 42u)];
+    OCR2A = sine_wave[(uint8_t)(sine_index[0] - 84u)];
 
 
 }
